# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MotiWish Backend API
  version: 0.2.0
  description: 面向 Android 端游戏化时间管理应用的统一后端接口。
paths:
...
  /api/v1/tasks/tasks/{id}/complete/:
    post:
      operationId: tasks_tasks_complete_create
      description: 完成任务实例以及应用奖励/惩罚。周期任务 settle_period=true 返回周期完成率、漏做次数以及惩罚明细。
      summary: 完成结算任务
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this 任务.
        required: true
      tags:
      - Tasks
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskAction'
            examples:
              完成今天任务:
                value:
                  progress: 100
              结算周期任务:
                value:
                  occurrence_date: '2026-06-04'
                  settle_period: true
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TaskAction'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TaskAction'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCompleteResponse'
          description: ''
  /api/v1/tasks/tasks/{id}/progress/:
    patch:
      operationId: tasks_tasks_progress_partial_update
      description: 更新任务实例进度，progress_target=目标值。
      summary: 更新任务进度
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this 任务.
        required: true
      tags:
      - Tasks
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedTaskProgressUpdate'
            examples:
              更新今天任务进度:
                value:
                  progress: 60
              更新日期任务进度:
                value:
                  occurrence_date: '2026-06-03'
                  progress: 80
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedTaskProgressUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedTaskProgressUpdate'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskProgressUpdateResponse'
          description: ''
  /api/v1/tasks/tasks/history/:
    get:
      operationId: tasks_tasks_history_retrieve
      description: 返回当前用户任务实例历史记录。
      summary: 获取任务历史记录
      tags:
      - Tasks
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskHistoryResponse'
          description: ''
  /api/v1/tasks/tasks/pricing/preview/:
    post:
      operationId: tasks_tasks_pricing_preview_create
      description: AI 定价会话创建前，后端归一化任务草稿，返回类型、结算轨道、难度估算、规模以及定价范围。
      summary: 预览任务定价上下文
      tags:
      - Tasks
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskPricingPreview'
            examples:
              探索轨道任务定价预览:
                value:
                  title: Debug training script
                  task_type: one_time
                  recurrence: none
                  settlement_track: exploration
                  auto_estimate_difficulty: true
                  estimated_focus_minutes: 180
                  tags:
                  - debug
                  - research
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TaskPricingPreview'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TaskPricingPreview'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskPricingPreviewResponse'
          description: ''
  /api/v1/tasks/tasks/today/:
    get:
      operationId: tasks_tasks_today_retrieve
      summary: 获取日期任务实例
      parameters:
      - in: query
        name: date
        schema:
          type: string
        description: 目标日期，格式 YYYY-MM-DD；默认今天。
      tags:
      - Tasks
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskTodayResponse'
          description: ''
...
components:
  schemas:
...
    DifficultyLevelEnum:
      enum:
      - low
      - medium
      - high
      type: string
      description: |-
        * `low` - 低
        * `medium` - 中
        * `high` - 高
...
    PatchedTaskProgressUpdate:
      type: object
      properties:
        occurrence_date:
          type: string
          format: date
          description: 任务实例日期，今天。
        progress:
          type: integer
          minimum: 0
          description: 任务实例实际进度，不能超过 progress_target。
...
    PricingStatusEnum:
      enum:
      - unpriced
      - pending
      - quoted
      - applied
      type: string
      description: |-
        * `unpriced` - 未定价
        * `pending` - 待 AI 定价
        * `quoted` - 已生成报价
        * `applied` - 已应用报价
...
    RecurrenceEnum:
      enum:
      - none
      - daily
      - weekly
      - monthly
      type: string
      description: |-
        * `none` - 不重复
        * `daily` - 每天
        * `weekly` - 每周
        * `monthly` - 每月
...
    SettlementTrackEnum:
      enum:
      - regular
      - exploration
      type: string
      description: |-
        * `regular` - 常规轨道
        * `exploration` - 探索轨道
...
    StatusCa7Enum:
      enum:
      - active
      - archived
      type: string
      description: |-
        * `active` - 启用
        * `archived` - 归档
...
    Task:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
          description: 任务标题。
          maxLength: 120
        description:
          type: string
          description: 任务说明，可为空。
        task_type:
          allOf:
          - $ref: '#/components/schemas/TaskTypeEnum'
          title: 任务类型
        recurrence:
          $ref: '#/components/schemas/RecurrenceEnum'
        settlement_track:
          allOf:
          - $ref: '#/components/schemas/SettlementTrackEnum'
          title: 结算轨道
        difficulty_level:
          allOf:
          - $ref: '#/components/schemas/DifficultyLevelEnum'
          title: 难度等级
        estimated_focus_minutes:
          type: integer
          minimum: 0
          nullable: true
          description: 探索轨道任务预计专注时长，分钟。
        weekdays:
          type: array
          items:
            type: integer
            maximum: 6
            minimum: 0
          description: 周任务使用，0 表示周一，6 表示周日。
        month_days:
          type: array
          items:
            type: integer
            maximum: 31
            minimum: 1
          description: 月任务使用，例如 [1, 15, 28]。
        metric_key:
          type: string
          description: 进度目标标识，可为空。
          maxLength: 50
        target_value:
          type: integer
          nullable: true
          description: 进度目标值，可为空。
        progress_target:
          type: integer
          maximum: 2147483647
          minimum: 0
          title: 进度目标
        reward_primary:
          type: integer
          maximum: 2147483647
          minimum: 0
          title: 一级货币奖励
        penalty_primary:
          type: integer
          maximum: 2147483647
          minimum: 0
          title: 一级货币惩罚
        pricing_status:
          allOf:
          - $ref: '#/components/schemas/PricingStatusEnum'
          readOnly: true
          title: 定价状态
        pricing_requested_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          title: 发起定价时间
        pricing_resolved_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          title: 完成定价时间
        pricing_snapshot:
          readOnly: true
          description: 后端定价上下文快照。
        starts_on:
          type: string
          format: date
          nullable: true
          title: 开始日期
        ends_on:
          type: string
          format: date
          nullable: true
          title: 结束日期
        due_at:
          type: string
          format: date-time
          nullable: true
          title: 截止时间
        status:
          $ref: '#/components/schemas/StatusCa7Enum'
        tags:
          title: 标签
        ai_metadata:
          title: AI 元数据
        created_at:
          type: string
          format: date-time
          readOnly: true
          title: 创建时间
        updated_at:
          type: string
          format: date-time
          readOnly: true
          title: 更新时间
      required:
      - created_at
      - id
      - pricing_requested_at
      - pricing_resolved_at
      - pricing_snapshot
      - pricing_status
      - task_type
      - title
      - updated_at
    TaskAction:
      type: object
      properties:
        occurrence_date:
          type: string
          format: date
          description: 任务实例日期，今天。
        progress:
          type: integer
          minimum: 0
          description: 当前进度值，探索轨道任务使用分钟。
        settle_period:
          type: boolean
          default: false
          description: 周期结算，true 返回惩罚明细。
    TaskCompleteResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/TaskOccurrence'
        success:
          type: boolean
        code:
          type: string
        message:
          type: string
      required:
      - code
      - data
      - message
      - success
...
    TaskHistoryResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TaskOccurrence'
        success:
          type: boolean
        code:
          type: string
        message:
          type: string
      required:
      - code
      - data
      - message
      - success
...
    TaskOccurrence:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        task:
          allOf:
          - $ref: '#/components/schemas/Task'
          readOnly: true
          description: 任务模板。
        occurrence_date:
          type: string
          format: date
          title: 发生日期
        status:
          $ref: '#/components/schemas/TaskOccurrenceStatusEnum'
        progress:
          type: integer
          maximum: 2147483647
          minimum: 0
          title: 完成进度
        settled_at:
          type: string
          format: date-time
          nullable: true
          title: 结算时间
        reward_transaction_id:
          type: integer
          nullable: true
          readOnly: true
        penalty_transaction_id:
          type: integer
          nullable: true
          readOnly: true
        settlement_details:
          readOnly: true
          description: 任务结算明细：奖励、惩罚、周期完成率。
        created_at:
          type: string
          format: date-time
          readOnly: true
          title: 创建时间
        updated_at:
          type: string
          format: date-time
          readOnly: true
          title: 更新时间
      required:
      - created_at
      - id
      - occurrence_date
      - penalty_transaction_id
      - reward_transaction_id
      - settlement_details
      - task
      - updated_at
    TaskOccurrenceStatusEnum:
      enum:
      - pending
      - completed
      - missed
      - cancelled
      type: string
      description: |-
        * `pending` - 待完成
        * `completed` - 已完成
        * `missed` - 已错过
        * `cancelled` - 已取消
...
    TaskPricingPreview:
      type: object
      properties:
        title:
          type: string
          default: ''
          description: 定价任务标题。
          maxLength: 120
        description:
          type: string
          default: ''
          description: 定价任务说明。
        task_type:
          allOf:
          - $ref: '#/components/schemas/TaskTypeEnum'
          description: |-
            任务类型。

            * `daily` - 日常任务
            * `recurring` - 周期任务
            * `one_time` - 一次性任务
        recurrence:
          allOf:
          - $ref: '#/components/schemas/RecurrenceEnum'
          default: none
          description: |-
            重复规则。

            * `none` - 不重复
            * `daily` - 每天
            * `weekly` - 每周
            * `monthly` - 每月
        settlement_track:
          allOf:
          - $ref: '#/components/schemas/SettlementTrackEnum'
          default: regular
          description: |-
            结算轨道: regular=常规轨道, exploration=探索轨道。

            * `regular` - 常规轨道
            * `exploration` - 探索轨道
        difficulty_level:
          allOf:
          - $ref: '#/components/schemas/DifficultyLevelEnum'
          default: medium
          description: |-
            任务难度，后端自动估算。

            * `low` - 低
            * `medium` - 中
            * `high` - 高
        auto_estimate_difficulty:
          type: boolean
          default: true
          description: 自动估算任务难度。
        estimated_focus_minutes:
          type: integer
          minimum: 0
          nullable: true
          description: 探索轨道任务预计专注时长，分钟。
        progress_target:
          type: integer
          minimum: 0
          default: 100
          description: 任务进度目标。
        metric_key:
          type: string
          default: ''
          description: 进度目标标识。
          maxLength: 50
        target_value:
          type: integer
          nullable: true
          description: 进度目标值。
        weekdays:
          type: array
          items:
            type: integer
            maximum: 6
            minimum: 0
          description: 周任务使用日期，0 表示周一。
        month_days:
          type: array
          items:
            type: integer
            maximum: 31
            minimum: 1
          description: 月任务使用日期。
        tags:
          type: array
          items:
            type: string
            maxLength: 50
          description: 任务标签。
      required:
      - task_type
    TaskPricingPreviewPayload:
      type: object
      properties:
        settlement_track:
          allOf:
          - $ref: '#/components/schemas/SettlementTrackEnum'
          description: |-
            结算轨道。

            * `regular` - 常规轨道
            * `exploration` - 探索轨道
        task_type:
          allOf:
          - $ref: '#/components/schemas/TaskTypeEnum'
          description: |-
            任务类型。

            * `daily` - 日常任务
            * `recurring` - 周期任务
            * `one_time` - 一次性任务
        formula:
          type: string
          description: 定价公式。
        normalized_payload:
          description: 后端任务定价上下文。
        difficulty_factor_hint:
          type: integer
          description: 探索轨道难度系数提示。
      required:
      - formula
      - normalized_payload
      - settlement_track
      - task_type
    TaskPricingPreviewResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/TaskPricingPreviewPayload'
        success:
          type: boolean
        code:
          type: string
        message:
          type: string
      required:
      - code
      - data
      - message
      - success
...
    TaskProgressUpdateResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/TaskOccurrence'
        success:
          type: boolean
        code:
          type: string
        message:
          type: string
      required:
      - code
      - data
      - message
      - success
...
    TaskTodayResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TaskOccurrence'
        success:
          type: boolean
        code:
          type: string
        message:
          type: string
      required:
      - code
      - data
      - message
      - success
    TaskTypeEnum:
      enum:
      - daily
      - recurring
      - one_time
      type: string
      description: |-
        * `daily` - 日常任务
        * `recurring` - 周期任务
        * `one_time` - 一次性任务
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT